+++ maia/utils/parallel/test/test_par_algo.py
@@ -2,6 +2,7 @@ import pytest_parallel
 
 import numpy as np
 
+from maia.utils import py_utils
 from maia.utils.parallel import algo as par_algo
 
 @pytest_parallel.mark.parallel(3)
@@ -40,3 +41,36 @@ def test_DistSorter(comm):
   sorter = par_algo.DistSorter(key, comm)
   assert (sorter.sort(data) == expected).all()
 
+
+@pytest_parallel.mark.parallel([1,4])
+def test_compute_gnum_s(comm):
+  # Simple test with strings  
+  keys = [['Pear', 'Banana', 'Peach', 'Banana'],
+          [],
+          ['Watermelon', 'Banana'],
+          ['Melon', 'Peach', 'Tomato']]
+  expected_gnum = [[5,1,4,1],
+                   [],
+                   [2,1],
+                   [6,4,3]]
+
+  if comm.Get_size() == 4:
+    rank_keys = keys[comm.Get_rank()]
+    rank_gnum = expected_gnum[comm.Get_rank()]
+  else: #Results should be // independant
+    rank_keys = py_utils.to_flat_list(keys)
+    rank_gnum = py_utils.to_flat_list(expected_gnum)
+
+  assert (par_algo.compute_gnum(rank_keys, comm) == rank_gnum).all()
+
+@pytest_parallel.mark.parallel(2)
+def test_compute_gnum_o(comm):
+  # Test with various objects
+  if comm.Get_rank() == 0:
+    keys = [test_compute_gnum_s, 'Banana', np.array([1,2,3], np.int32)]
+    expected_gnum = [5,1,2]
+  elif comm.Get_rank() == 1:
+    keys = [None, False, test_compute_gnum_s, 'Banana']
+    expected_gnum = [4,3,5,1]
+
+  assert (par_algo.compute_gnum(keys, comm) == expected_gnum).all()